Only prepend glyph classes with "@" if not already
Glyphs allows classes to start with "@", and seems to treat them the
same as those that don't. It even exports UFOs with class names
starting with "@@", which is illegal syntax.

For glyphs2ufo, let's always make sure the class names are legal, so
that all classes should be treated the same by any program using the
output UFOs.