Remove conflicting kerning rules as they're added
This way, the UFOs output by the library will always have legal
kerning. This is important for interpolation because MutatorMath uses
defcon which requires it.

Don't print reports from defcon exceptions, since they are only used
for conflicting kerning (which shouldn't happen) and in other cases
catching and re-raising an exception makes it harder to see where the
error actually came from.