[interpolation] Always generate designspace path
With this, the interpolate and build_designspace functions always
create their own path to use when saving the designspace document.

It would be nice to just use the glyphs path, replacing the extension,
but perhaps the glyphs path can't always be trusted. For example
fontmake creates a temporary glyphs path when preprocessing the glyphs
source contents.