Use Glyphs' custom style axis
This adds a third axis for interpolation which can be used in Glyphs:
https://www.glyphsapp.com/tutorials/multiple-masters-part-1-setting-up-masters

As a result we need to be slightly more flexible in how we parse style
names. We also no longer try to copy features, groups and info to
instances from a Regular font, since such a font may not exist.
Instead, we just copy from the first given master.