Remove `italic` argument
This was used to specify that all masters and instances in a single
Glyphs source were italic, but it may not be safe to assume that's
always the case.

Now fonts are considered italic if their italicAngle attribute is
nonzero, where before the italic argument was generally determined by
checking if "Italic" was in the source path. I think this new
heuristic is better.