Add code for writing typed/untyped data to glyphs files.
This changes the parsing code to handle octal/unicode escapes as well
as quotes.  This means that unicode values are returned from the parser,
and may be passed to the writer.

The casting code no longer normalizes curly quotes, they are maintained in
the data.  If they need to normalized clients will need to do that conversion.
This allows for better round-tripping of the original glyphs data.

Note that it is not possible to fully round-trip the glyphs source.  In
particular leading spaces are not preserved.  Quoting of strings is also
not fully preserved, instead the determination of whether strings need to
be quoted is done based on the contents of the string.  Glyphs seems to
quote and indent based on the position of the element in the hierarchy and
not on the content.  This does not appear to impact Glyphs' ability to
load the rewritten source.

Analysis of the casting code revealed that the dictionaries describing the
type structure were never modified and so did not need to be created/copied,
so the former functions returning them were changed to constants.

The original casting functions were converted to read methods of a collection
of objects, to which corresponding 'write' methods were added, and the
loop that iterated over the structure was changed to call a 'convert' method
with a parameter indicating whether to read or write.

In some cases the original casting functions were cloning parts of the
data structure they were operating on, and some of these were changed to
modify the data in place like the bulk of the code.  This was not done
in all cases, however-- in particular map is used in a number of functions
and it created a new list (which is then substituted for the original).