Later kerning pairs take precendence in conflicts
With this change, kerning conflicts are checked and resolved in
reverse order of how they appear in Glyphs source. This is to try and
produce kerning in the same way as Glyphs itself, where the last
instance of a pair is used to create its rule (while this code drops
any rule for a pair *already* seen).