builder.py: for closed contours, always move the last node to first in UFO .glif
For closed contours, Glyphs.app stores the starting node (as indicated by the position of the blue triangle icon) at the end of the nodes list. In the UFO we need to put it at the beginning, otherwise the position of the start point changes.

Should fix https://github.com/googlei18n/fontmake/issues/142#issuecomment-246337019