Expose more functionality for interpolating
This returns a designspace path and instance data when a designspace
is generated in the build_master function, and adds a function for
incorporating this instance data (which reflects the structure of
Glyphs source) into actual UFOs.

Ideally I'd like to take out interpolation from glyphsLib completely,
and only have options for returning master UFOs and a designspace. But
for now I'm not sure if anyone is using the interpolation feature.