[interpolation] Copy lib data to instances
Some of this data will be technically incorrect for most instances,
for example the weight name and value. However, all of this incorrect
data resides in entries with a Glyphs-specific prefix and will
generally be ignored.

What we care about is the public.glyphOrder entry, which does need to
be present in any generated instance (and should be the same across a
family).