Explicitly set OpenType ascender/descender fields
These turn out to be pretty complicated and we can't trust they'll
be set correctly by a UFO compiler just from the generic fields.

Glyphs does something a little different, but this is a simple
treatment which yields predictable results for the common case where
every master has the same ascender and descender values.