.travis.yml: also build on tag "branches"; install from requirements.txt
When pushing a tag, Travis sees that as a different 'branch'. So we need to enable builds on 'branches' whose name matches a version string, e.g. 'v1.1.0'.

It's important to not forget the 'v' prefix when creating a tag, otherwise Travis will skip the build (and hence the deployment).

(that's what tools like bumpversion help you with, but don't want to over-complicate this for now)