Do not generate inactive instances
Before this change, the code was looking for an attribute named
`active`, but this is not what Glyphs.app stores in its files when an
instance gets marked as inactive in the Glyphs user interface. Instead,
Glyphs calls this attribute `exports` in its files.

Fixes https://github.com/googlei18n/glyphsLib/issues/129.