Re-generate the built-in Glyphs data
The upstream repo has had a few changes on April 15, 20017, but these
only affect the upstream tooling and not any actual data. (I've gone
through each upstream change to confirm this). Therefore, this patch
only updates the upstream commit hash without changing the embedded
data. So it's a no-op, but it still seems useful to clarify that we're
not lagging behind upstream.