[casting] add 'exports' as (truthy) instance parameter
The default value for an instance's 'exports' paramters is '1', i.e.
all instances are exported by default, and noramlly there's no explicit
line in the .glyphs file that says 'exports = 1', as it's implied.

One can disable generating an instance by unticking the checkbox
'is active'; in which case, Glyphs.app will add an 'exports = 0'
parameter to the instance data.

The instances' 'exports' parameter must be read in as a boolean (True
or False) value, and not as simple string.