[interpolation_test] only set exports=False (True is implied); use bool instead of int
Glyphs.app never writes 'exports = 1' nor 'active = 1' in the .glyphs source
file, as it is the default value and it's implied.

Similarly, in the tests, we only need to override the parameter when
it's value is not the default, e.g. False.

We use bool instead of int here, because both 'exports' and 'active'
parameters are being cast to bool type on reading the source data.