Refactor how glyphsLib finds the "Regular" master
In MutatorMath, one of the masters is used as the base for copying
default font information, and glyphsLib tries to use the "Regular"
master for this.

Before this change, the logic for finding the "regular" master was
interleaved with writing out the masters into the output designspace
file for MutatorMath. After this change, the logic is in a separate
function.

No change in functionality, just refactoring.