[builder] populate kerning groups following the glyphOrder custom parameter, if present
When a 'glyphOrder' custom parameter is present, we load the the glyphs data (including left/right kerning group membership) in that order, otherwise we keep using the glyphs data's intrinsic order.

Kerning groups are stored as arrays of strings in the UFO groups.plist, whereas they are attributes of the glyph objects in Glyphs.app. Thus, the original ordering of the glyphs inside the UFO kerning groups is lost if these were to be imported in a glyphs file.

Although the order is not relevant to make a kern feature, it's important if one wishes to roundtrip UFO groups.plist -> Glyphs' {left,right}KerningGroup -> UFO groups.plist, and avoid producing spurious changes in the diff.

By using the glyphOrder, we can try to ensure a more predictable order, provided that the members of the UFO kerning groups were sorted by 'public.glyphOrder' to begin with.