Remove shebangs from non-executable modules
Since these files are not supposed to be executable, there is not point
in having shebangs there. Also, since they just point to the system
python binary, they end up pointing to the wrong python version if both
python2 an python3 packages are installed for one of the python versions.