[builder] don't set preferred family/subfamily; add TODO on style linking
There is no point setting these optional fontinfo attributes if they
are the same as the default UFO familyName and styleName.

Added TODO about isBold, isItalic and linkStyle properties.
The classtree branch already supports parsing these.
We should use them to generate proper styleMap* names and write these
in the designspace, so that mutatorMath adds them to the instances.