[interpolation] remove call to set_redundant_data and set_default_params
We need more specific functions for instances, than a generic
`set_redundant_data` that applies to both masters and instances.

The `set_defaults_params` is completely uselss in the context of
instances, since the attributes it is trying to set are already
set on instances, because the same function was called on the masters
already, and the instances would have inherited those default values.
I might add a test for that later.