[builder] revert to setting basic styleMap* names for masters
It's not clear what the fallback value for a missing `styleMapFamilyName` should be, the UFO spec doesn't say.

E.g. ufo2ft.fontInfoData simply concatenates preferred family and preferred subfamily; however, this leads to a "Regular" master TTF that is going to be used with varLib to create a variable font to have nameID=1 like "Noto Sans Regular" instead of "Noto Sans", and that's not good.

It's better we provide basic styleMapFamilyName and styleMapStyleName for the master UFOs which make sense, even if we don't have the full style linking information as we do for instances.

Also, this moves build_stylemap_names function from interpolation to builder module.