Skip dangling layers and layers without a name on UFO building
This change detects and skips layers that are not associated with an
existing master and layers without a name. The former are a sort of data
corruption. Empty layer names are an error as per the UFO specification
and choke ufoLib. Both may happen when you copy glyphs from a different
font in Glyphs, Glyphs simply ignores them.

All invalid layers now get a warning printed and skipped. Before, they
could have been silently ignored.

Add tests for both types of bogus layers.