Improve color tuple handling
Glyphs saves a glyphs' color tuple as a List of type [u8, u8, u8, 1]. The alpha
channel is not supported as of Glyphs 2.5.1 (confirmed by Georg Seifert by
mail) and is always set to 1. The UFO specification uses floats to store color
values, using 3 decimal points is enough to round-trip u8 to float and back.