the default designspace.filename should use the provided custom 'family_name'
this looks like a regression from glyphsLib 2.2.1.
Previously the returned 'designspace_path' from the glyphsLib.builder_masters function
that was used in fontmake<=1.4 would return a filename that included the custom
--family-name (when provided); in glyphsLib 2.3, the designspace.filename is always
set to the GSFont.familyName, regardless of whether the custom --family-name option was
passed.

This is a problem because fontmake by default uses the same name as the designspace file
as the output filename for the generated variable font.
When building, in sequence and to the same variable_ttf output directory, multiple
Noto fonts (e.g. NotoSansBengali and NotoSansBengaliUI) that only differ from each other
because of that --family-name option (they share the same .glyphs source, only differ
for the --mti-source .plist file to build the opentype features), the result is that
the second will the overwrite the first one, as the generated designspace files will
overwrite each other, likewise the output variable fonts.

Anyway, it's a regression and this fixes it.