add generate_GDEF argument, independent of minimize_glyphs_diffs
The default value of --generate-GDEF option for the console script is False, to keep
it in line with the rest of the options that are more geared towards exporting UFO
sources with the intent of roundtripping back to Glyphs.

For the programmatic interface (build_masters, to_designspace, UFOBuilder) the default
value for generate_GDEF is True, again to not break previous assumptions where
the minimize_glyphs_diffs argument was already set to False, and the propagate_anchors
also defaults to True. That is, these functions's defaults are more geared towards
generating UFOs as temporary build artifacts for the sake of building a binary OTF
in the fontmake pipeline.

In the future we may want to revise these incoherent default values between the cli
and the API.

Also, we currently bail out with a ValueError if the features already
contain a GDEF table. The user needs to remove it manually or call with
the generate_GDEF argument set to False.