remove unicodes list hack; write unicode with quotes if looks like decimal number
multiple unicodes are always quoted now in current Glyphs.app, we don't need to keep
supporting the (invalid) multiple comma-separated unicodes that aren't wrapped in quotes.

We also write single unicode within quotes to make it clear that these are not to be
interpreted as decimal integers but as (hexadecimal) strings.

Unless the unicode string starts with a non-digit, in which case it cannot be a decimal
number, so the quotes can be omitted.