set KernFeatureWriter with mode="append" in designspace lib
The default mode for ufo2ft kernFeatureWriter is going to change
back to "skip", as that's more common and expected in the UFO world.
Therefore to keep the current behavior (append autogenerated kerning
to any user-defined lookup, usually used for contextual kerning), we
need to tell ufo2ft that we want "append" mode instead.

We store this information in the designspace lib element. Fontmake
will take care of passing that on to ufo2ft compiler, when building
from a designspace that was generated by glyphsLib (or any designspace
actually).