Support public.skipExportGlyphs lib key for storing export flags (#519)
On glyphs2ufo, this stores Glyphs.app's export flags in the Designspace and UFO-level "public.skipExportGlyphs" lib key but does not write the now-obsolete glyph level "com.schriftgestaltung.Glyphs.Export". On ufo2glyphs, it reads the Designspace level "public.skipExportGlyphs" and glyph level "com.schriftgestaltung.Glyphs.Export". When going ufo2glyphs from a bunch of UFOs, it will read the UFO-level "public.skipExportGlyphs" lib key but expect it to be the same or not exist in all UFOs.