custom_params: register ReplacePrefix before ReplaceFeature
This way, ReplacePrefix will be applied before ReplaceFeature.
This is important because, when these two params are used in combination
and the replaced feature contains a reference to a lookup only contained
in the prefix to be replaced, feaLib Parser will trigger an error if
it cannot resolve the lookup reference. Thus it is important to replace
the prefix before the feature body.