setup.cfg: fonttools[ufo] extra is required by defcon
If one installs glyphsLib in empty venv, the 'fs' requirement will not be
fetched automatically and running glyphs2ufo fails with ImportError.
This is because we specify defcon _and_ fonttools in install_requires,
and pip dep resolver isn't clever enough...
Mostly people install glyphsLib via fontmake (which requires fonttools
with the [ufo] extra) so they don't realize about this issue.