adapt code to be compatible with either defcon or ufoLib2 APIs
ufoLib2 objects do not have backreferences to their parents
so we need to get these differently or pass them down explicitly.

postscriptBlueValues and postscriptOtherBlues are None by default
in ufoLib2, whereas they are empty lists in defcon.

Glyph.markColor is a plain string in ufoLib2, not a smart Color
object.