Keep Axis axes (#595)
Make sure here that if there is an Axes parameter, we actually really
keep axes that are defined in it, even if they do nothing. This
prevents an edge case where you have a single master at the default
location that ends in with a Designspace with no axes, as they were
all filtered out.