ci/deploy: no need strip first line of tag message
When we were using Travis to create the Github release the first line of a tag's message was used as the release name, so it was supposed to repeat the tag version. We don't need that now.
The whole tag message (except the optional PGP signature) **is** the release notes's body.