[builder] Fix color on master layer optimization
When a master layer has color attributes, we try to avoid creating a
duplicate color layer of it. This was broken if the layer had multiple
paths with with different color attributes each. Fix this by returning
the layer only when the new layer has the same number of paths and
components, and create a new layer otherwise.