add option to both UFO|GlyphsBuilders to expand_includes in features
As suggested in https://github.com/googlefonts/glyphsLib/issues/797, it may be useful sometimes to resolve and expand include statements in features, to create self-contained UFOs/glyphs sources that don't reference any external files. It is disabled by default, to make sure rountripping continues to work. It is not conflated with 'minimal' option because they are distinct things, one may want to enable one without the other.