Fix handling of COLRv0 layers
Glyphs handles components for color layers in a special way. If the
component glyph has color layers of its own, the component use the first
color layer with the same color index, otherwise it fallback to the
default layer. We try to do that here as well.

Fixes https://github.com/googlefonts/glyphsLib/issues/866