Remove very slow code that does nothing
It is iterating over all masters in the font, then tries to access the
master by its id and then checks if the master is None, but this can
never happen since we are iterating over already existing masters.

This code is called in a very hot spot, and in a script I’m profiling,
it takes more than 50% of the font build time.

Fixes https://github.com/googlefonts/glyphsLib/issues/646