[axes_test] test that axis min/max are influenced by virtual masters params
The test font 'IntermediateLayer.glyphs' has a couple of glyphs that use virtual masters to control the cap height via a 'CPHT' axis.
It contains no Axis Mappings nor Axis Location parameters, the axis coordinates are intented to be un-mapped (user==design). The virtual master custom parameters are expected to be taken into account when computing the CPHT min and max values.