invert anchor order to trigger anchor_test.py fail
In anchor_propagation.py, we have code to adjust propagated anchor position when a mark component is anchored to a specific named anchor as specified in the ComponentInfo lib key.
The code indiscriminately applies the adjustment in a for loop, without checking the anchor names match the expected component anchor.
I modified one of the AnchorPropagation.glyphs test file by changing the order of anchors in a mark glyph used as component in order to trigger a failure. Will follow up with a fix