Store MATH Glyphs glyph.userData in UFO glyph userData
The Glyphs MATH plugin makes distinction between userData that does not
change per master and stores them in glyph.userDara, and that changes
per-master and stores then in layer.userData. When writing to UFO we
were maintaining this distinction by keeping the former in font.lib and
the layer in glyph.lib.

But since a UFO file represent a single master, this distinction makes
no much sense, and the UFO data looks split arbitrarily. We now keep
both glyph and layer userData in glyph.lib. When converting from UFO to
Glyphs if the glyph-level data is different between UFO masters, we
issue a warning.