[anchor_propagation] do propagate if comprises only 1 mark component
if a composite glyph (e.g. 'cedilla') has exactly one component which contains any '_'-prefixed anchors (i.e. has a 'mark component' in the terminology used in this anchor_propagation.py), all of the component's anchors should be propagated to the parent composite glyph; and not only (as it was currently the case) when the parent's name resembles a ligature (e.g. 'shadda_fatha-ar').
This is an attempt to converge towards https://github.com/googlefonts/fontc/issues/811