markFeatureWriter: Optimize contextual anchor lookups
Instead of making one `pos` statement per individual base glyph (which
translates to one subtable, at least in feaLib), group all base glyphs
in one statement. When there is a large number of base glyphs with the
same contextual anchor, this can save quite a bit (for a font I’m
working on, this save ~4kb for an anchor with 93 base glyphs).