demote logging.warning about missing component layer to debug level
The current propagate_anchors implementation only finds a matching component layer that has the same layerId so only works for 'master' layers; to reduce the verbosity of the log for fonts that contain lots of special non-master (brace, bracket, etc) layers, for which we currently can't retrieve the corresponding component layer's anchors, for now we demote this from a warning to a debug message.
Note that the old anchor_propagation.py was only processing 'master' layers as well, so ignoring the special layers during anchor propagation is not a regression.