Support fonts named Oblique instead of Italic (#1034)
We have a font that is an Oblique, not a true Italic, so we named it Oblique. When creating UFOs, or building static TTFs, the style name for the Obliques would be "Oblique Italic" and "Bold Oblique Italic". I added "Oblique" as an alternative where names were previously tested only for "Italic".

While investigating this, I noticed that the styleMapStyleName was not getting set correctly in the case of Bold Italic (it was getting set as Italic, not Bold and Italic), thus the macStyle and fsSelection bits were not getting set correctly (both the Italic and Bold Italic had bits set the same: i.e. Italic and not Bold). So I changed is_bold=(stylename=="Bold") to is_bold=("Bold" in styleName) in to_ufo_names.

I also added test cases for both.

---------

Co-authored-by: Rob McKaughan <robmck@microsoft.com>