Fix extra segment in eraseOpenCorners
In the process of porting this functionality to Rust, I've run into a
funny issue: in some cases, after erasing corners, the outline generated
here contained an extra duplicate point at the beginning.

This error seems to be introduced only for contours where the last
segment is a curve, and it is introduced early; even before we start
processing, there is an extra segment in the segments list for these
contours.

This patch addresses this by changing the logic of closePath to not emit
a final line segment if it would be a duplicate, but I'm not totally
confident that this is the right approach.