Add additional license text for GlyphData
The main `LICENSE` in https://github.com/schriftgestalt/GlyphsInfo/ is
MIT and is already reproduced here as `GlyphData_LICENSE`; however, the
`README.md` for `GlyphsInfo` attributes the data in part to the Adobe
AGL, with BSD-3-Clause terms. This commit adds that license text to a
new file `GlyphData_AGL_LICENSE`.