glyphdata: strip unused script -suffix from ligature production name
E.g. a ligature glyph named "one_one-han.paren" (as found in ButTaiwan/iansui) gets renamed "one_one.paren" in Glyphs.app whereas in glyphsLib it was keeping the original name including the script suffix (hyphen stripped) but the script itself is not used by any of the ligature components.