Sort bracket layers by glyph id
Bracket layers are added in two phases: first we add layers as they
exist in glyphs, in the order that glyphs are listed in the font; and
then later on we go and add layers for any component glyphs that have
components with bracket layers.

This means the layers end up in a funny order, and then this order can
impact the rules that we generate in the final font.

This tries to standardize this, by sorting the layers by glyph order
after the components are added.

At a higher level I find it odd that order matters at all here, and I
wonder if there could be a better approach? But for the time being this
will at least give us something reasonable that fontc can replicate.