Skip irrelevant layers when converting brackets
This is a funny little bug: when converting brackets to designspace
rules, we reuse the 'overlayFeatureVariations' function from varLib to
create a set of minimal designspace regions to represent the set of of
possible substitutions.

overlayFeatureVariations has logic for clearing out axes that are
irrelevant (that is, where a rule would match across an entire axis) but
this logic expects that the regions are expressed in normalized
coordinates. When it is called form glyphsLib, the regions are expressed
in userspace coordinates, and this logic is never triggered.

To work around this, we manually skip axis rules that match the entire
axis before calling overlayFeatureVariations.